wip - moving messaging logic
